+++ CHANGELOG.md
@@ -5,6 +5,16 @@ This file serves to keep track of changes across versions of the app.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+
+- Ripple effect when event is clicked in Schedule view
+
+### Fixed
+
+- Quotation marks properly rendering in one of the events
+
 ## [1.0.2] - 2022-04-08
 
 ### Fixed
@@ -155,7 +165,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Images not properly loading from cache
 - Showing _alt_ text in case images don't load
 
-[Unreleased]: https://github.com/biocienciasgrx/ceeebi/compare/v0.6.1-alpha...HEAD
+[unreleased]: https://github.com/biocienciasgrx/ceeebi/compare/v0.6.1-alpha...HEAD
 [1.0.2]: https://github.com/biocienciasgrx/ceebi/releases/tag/v1.0.2
 [0.8.8-beta]: https://github.com/biocienciasgrx/ceebi/releases/tag/v0.8.8-beta
 [0.8.7-beta]: https://github.com/biocienciasgrx/ceebi/releases/tag/v0.8.7-beta